[frontend] refactor useHasOnlyAccessToImportDraftTab to useImportAccess
