+++ opencti-platform/opencti-front/src/private/components/drafts/Drafts.tsx
@@ -17,7 +17,7 @@ import { computeValidationProgress } from '../../../utils/draft/draftUtils';
 import { addFilter, emptyFilterGroup, useBuildEntityTypeBasedFilterContext } from '../../../utils/filters/filtersUtils';
 import useConnectedDocumentModifier from '../../../utils/hooks/useConnectedDocumentModifier';
 import useDraftContext from '../../../utils/hooks/useDraftContext';
-import useHasOnlyAccessToImportDraftTab from '../../../utils/hooks/useHasOnlyAccessToImportDraftTab';
+import useImportAccess from '../../../utils/hooks/useImportAccess';
 import { usePaginationLocalStorage } from '../../../utils/hooks/useLocalStorage';
 import { UsePreloadedPaginationFragment } from '../../../utils/hooks/usePreloadedPaginationFragment';
 import useQueryLoading from '../../../utils/hooks/useQueryLoading';
@@ -129,7 +129,7 @@ const Drafts: FunctionComponent<DraftsProps> = ({ entityId, openCreate, setOpenC
   const draftColor = getDraftModeColor(theme);
   const validatedDraftColor = theme.palette.success.main;
   const draftContext = useDraftContext();
-  const hasOnlyAccessToImportDraftTab = useHasOnlyAccessToImportDraftTab();
+  const { hasOnlyAccessToImportDraftTab } = useImportAccess();
   const { setTitle } = useConnectedDocumentModifier();
   if (!entityId) {
     setTitle(t_i18n('Drafts'));
